+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Alliance Ledger show Corporations instead of Characters
 - Renamed template to information
 - Logger System
+- ETAG System use log timing instead of own timing log
 
 ### Removed