+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - AA min. requirements from 4.0.0 to 4.6.0
 - Refactor Corpjournal Manager
 - Refactor Template System
+- Refactor Django Templates
 - Alliance Ledger show Corporations instead of Characters
 - Renamed template to information
 - Logger System